requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple/
altair==4.2.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0; python_version >= '3.6'
astor==0.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
backports.zoneinfo==0.2.1; python_version < '3.9'
base58==2.1.1; python_version >= '3.5'
beautifulsoup4==4.10.0; python_version >= '3.1'
biopython==1.77
bleach==4.1.0; python_version >= '3.6'
blinker==1.4
boto3==1.20.41; python_version >= '3.6'
botocore==1.23.41; python_version >= '3.6'
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10; python_version >= '3'
click==8.0.3; python_version >= '3.6'
covid==2.3.0
cycler==0.11.0; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints==0.3; python_version >= '2.7'
enum-compat==0.0.3
idna==3.3; python_version >= '3'
importlib-resources==5.4.0; python_version < '3.9'
ipykernel==5.5.6; python_version >= '3.4'
ipython-genutils==0.2.0
ipython==7.16.3
ipywidgets==7.6.5
jedi==0.17.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jinja2==3.0.3; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==4.4.0; python_version >= '3.7'
jupyter-client==7.1.1; python_full_version >= '3.6.1'
jupyter-core==4.9.1; python_version >= '3.6'
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2; python_version >= '3.6'
kiwisolver==1.3.2; python_version >= '3.7'
markupsafe==2.0.1; python_version >= '3.6'
matplotlib==3.2.2
mistune==0.8.4
nbclient==0.5.10; python_version >= '3.7'
nbconvert==6.4.0; python_version >= '3.7'
nbformat==5.1.3; python_version >= '3.5'
nest-asyncio==1.5.4; python_version >= '3.5'
notebook==6.4.7; python_version >= '3.6'
numpy==1.19.0
packaging==21.3; python_version >= '3.6'
pandas==1.0.5
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.0.0; python_version >= '3.7'
prometheus-client==0.12.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prompt-toolkit==3.0.24; python_full_version >= '3.6.2'
protobuf==3.19.3; python_version >= '3.5'
ptyprocess==0.7.0; os_name != 'nt'
pyarrow==6.0.1; python_version >= '3.6'
pycparser==2.21
pydantic==1.9.0; python_full_version >= '3.6.1'
pydeck==0.7.1; python_version >= '3.7'
pygments==2.11.2; python_version >= '3.5'
pyparsing==3.0.7; python_version >= '3.6'
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2021.3
pyzmq==22.3.0; python_version >= '3.6'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
s3transfer==0.5.0; python_version >= '3.6'
send2trash==1.8.0
setuptools==60.5.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.1; python_version >= '3.6'
streamlit==0.63.0
terminado==0.12.1; python_version >= '3.6'
testpath==0.5.0; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.11.2; python_version >= '3.5'
tornado==6.1; python_version >= '3.5'
traitlets==5.1.1; python_version >= '3.7'
typing-extensions==4.0.1; python_version >= '3.6'
tzdata==2021.5; python_version >= '3.6'
tzlocal==4.1; python_version >= '3.6'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
validators==0.18.2; python_version >= '3.4'
watchdog==2.1.6; python_version >= '3.6'
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0; python_version < '3.10'
